Update patch set 1
Patch Set 1:

Hi Koren,

You are going to need to run this in a venv, I tried to shim it in for you but got stuck at pip3 install mysqlclient as mysqlclient does not support python3. If you look at the bottom of  relng/jjb/opnfvdocs/docs-rtd.yaml you can see how this could be accomplished.

Patch-set: 1